Updated form validation - work in progress; cleaned up some CSS and calculations

+++ src/account/account-list/account-list-controller.js
@@ -88,7 +88,9 @@ var accountListController = function( $scope, AccountFactory ) {
      */
 
 	$scope.deleteAccount = function( account ) {
-		AccountFactory.deleteAccount ( account );
+          if ( confirm( 'Are you sure you want to delete this account?') ) {
+		   AccountFactory.deleteAccount ( account );
+          }
 	};
 
 	/**
